% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Plots.R
\name{Bior_DotPlot}
\alias{Bior_DotPlot}
\title{Dot Plot}
\usage{
Bior_DotPlot(
  x,
  y,
  size,
  group.by,
  colour = NULL,
  max_size = 5,
  text.size = 15,
  breaks = waiver(),
  theme = theme_bw(),
  labs.x = "",
  labs.y = "",
  title = "",
  legendtitle.color = "Types",
  legendtitle.size = "Value"
)
}
\arguments{
\item{x}{A vector, x-axis value}

\item{y}{A vector, y-axis value}

\item{size}{A vector, point size value}

\item{group.by}{color group by x or y}

\item{colour}{colour}

\item{max_size}{max point size}

\item{text.size}{text size}

\item{breaks}{scale_size_area(breaks)}

\item{theme}{choose ggthemes, default:theme_bw()}

\item{labs.x}{labs x}

\item{labs.y}{labs y}

\item{title}{title}

\item{legendtitle.color}{guides(color=guide_legend(legendtitle.color))}

\item{legendtitle.size}{guides(size=guide_legend(legendtitle.size))}
}
\value{
A ggplot object
}
\description{
Create a dot plot.
}
\examples{
# Examples 1
x <- rep(paste('sample', 1:5, sep=''), 4)
y <- rep(paste('gene', 1:4, sep=''), 5)
size <- round(rnorm(20, mean = 10, sd = 5))
colour <- c("#1F77B4FF","#FF7F0EFF","#2CA02CFF","#D62728FF","#9467BDFF")
p <- Bior_DotPlot(x = x, y = y, size = size, group.by = x, colour = colour, max_size=10)
p
}
