% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Plots.R
\name{Bior_StackBarplot}
\alias{Bior_StackBarplot}
\title{Stack Barplot}
\usage{
Bior_StackBarplot(
  data,
  x.order = NULL,
  type.order = NULL,
  col = scales::hue_pal(),
  label.size = 5,
  labs.x = "",
  labs.y = "",
  title = "",
  legend.key.size = 1,
  text.size = 15,
  bar.width = 0.7,
  theme = theme_bw()
)
}
\arguments{
\item{data}{A dataframe, the columns are x-axis sample, the rows are fill type}

\item{x.order}{x axis sample order}

\item{type.order}{fill type order}

\item{col}{fill colours}

\item{label.size}{label size}

\item{labs.x}{x axis title}

\item{labs.y}{y axis title}

\item{title}{figure title}

\item{legend.key.size}{legend size}

\item{text.size}{text size}

\item{bar.width}{bar width}

\item{theme}{choose ggthemes, eg:theme_bw(), theme_classic()}
}
\value{
A ggplot object
}
\description{
Create a stacked barplot.
}
\examples{
# Examples 1
data <- data.frame(matrix(rnorm(20, mean = 20, sd = 5),c(4,5)))
data <- round(data)
rownames(data) <- paste('gene',1:4, sep='')
colnames(data) <- paste('sample',1:5,sep='')
x.order <- c('sample1','sample5','sample4','sample3','sample2')
type.order <- c('gene1','gene2','gene3','gene4')
col <- c("#AEC7E8FF","#FFBB78FF","#98DF8AFF","#FF9896FF")
p <- Bior_StackBarplot(data, x.order=x.order, type.order=type.order, col = col,
                       labs.x='Samples', labs.y='Numbers', title='Test Bior_StackBarplot',
                       theme=theme_minimal())
p
}
