% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Plots.R
\name{Bior_PiePlot}
\alias{Bior_PiePlot}
\title{Pie Plot}
\usage{
Bior_PiePlot(
  x,
  labels,
  col = pal_d3("category20,", alpha = 0.7)(20),
  title = "",
  fontsize = 20,
  legend.key.size = 1
)
}
\arguments{
\item{x}{A vector of value}

\item{labels}{A vector of labels for x}

\item{col}{colour}

\item{title}{title}

\item{fontsize}{fontsize}

\item{legend.key.size}{legend size}
}
\value{
A ggplot object
}
\description{
Create a Pie plot.
}
\examples{
# Examples 1
x <- c(3,7,9,1,2)
labels <- c("A", "B", "C", "D", "E")
col <- c("#AEC7E8B2", "#FFBB78B2", "#98DF8AB2", "#FF9896B2", "#C5B0D5B2")
p <- Bior_PiePlot(x, labels, col=col, title="Test Bior_PiePlot")
p
}
