% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Plots.R
\name{Bior_Sankeyplot}
\alias{Bior_Sankeyplot}
\title{Sankey Plot}
\usage{
Bior_Sankeyplot(
  links,
  Nodes.colour = NULL,
  Nodes.order = NULL,
  fontSize = 12,
  nodeWidth = 30,
  nodePadding = 10,
  margin = NULL,
  height = 600,
  width = 600,
  sinksRight = TRUE
)
}
\arguments{
\item{links}{A dataframe, colnames must have ‘source’ ‘target’ ‘value’}

\item{Nodes.colour}{Set nodes colour}

\item{Nodes.order}{Set nodes order}

\item{fontSize}{Set fontsize}

\item{nodeWidth}{Set nodewidth}

\item{nodePadding}{Set the gap size between nodes}

\item{margin}{R margin}

\item{height}{numeric height for the network graph’s frame area in pixels}

\item{width}{numeric width for the network graph’s frame area in pixels}

\item{sinksRight}{boolean. If TRUE, the last nodes are moved to the right border of the plot}
}
\description{
Create a sankey plot.
}
\examples{
# Examples 1
# links data, colnames must have 'source' 'target' 'value'
links <- data.frame(
  source=c("C","A", "B", "E", "D"),
  target=c("b","c", "a", "e", "d"),
  value=c(1, 2, 0, 4, 5)
)

# Set Nodes order and colour
Nodes.order <- c("A", "B", "C", "D", "E", "a", "b", "c", "d", "e")
Nodes.colour <- c("#1F77B4B2","#FF7F0EB2","#2CA02CB2","#D62728B2","#9467BDB2",
                  "#8C564BB2","#E377C2B2","#7F7F7FB2","#BCBD22B2","#17BECFB2")
p <- Bior_Sankeyplot(links, Nodes.order=Nodes.order, Nodes.colour=Nodes.colour, fontSize=20)
p
}
